daos/ItemOptions: Do not default-initialize `@readonly` properties
PHP does not allow specifying an explicit default value on readonly properties and PHPStan 2.0 enforces the same for `@readonly` PHPDoc marker.

Let’s move the initial values to else branches. This required turning the method into a constructor.

Also fix `@readonly` for `$extraIds` property – it needs to be on a separate line for PHPStan to recognize it.