client: Switch `Route`s from `children` to `render` props
In react-router 5, when a function is passed to `children` prop,
it is rendered even when the `Route`’s `path` does not match.
We did that with `EntriesPage`.

https://v5.reactrouter.com/web/api/Route/route-render-methods

Since we do not need to animate `EntriesPage` or anything,
rendering it when not on `/entries` page is pointless.
Let’s switch it to `render` prop.

The upcoming react-router 6 migration will require all routes to use
the `element` prop so, while at it, let’s also switch the other routes
to make the future diff cleaner.