client: Fix CheckAuthorization redirection
The `CheckAuthorization` was no longer redirecting to the `/sign/in` page and the following warning was logged into the console:

> You should call navigate() in a React.useEffect(), not when your component is first rendered.

Probably regressed in 88deedea994efe4cdf97207deed9a292057da80a.

Let’s fix the redirection.

Using polyfill for `useEffectEvent` from react 1.19.2 to get `eslint-plugin-react-hooks` of our back.

Keeping the error message in the body to have some content in case the effect is slow.