Controllers\Sources: Remove unused `GET /source` API endpoint
Initially, it was used to produce HTML fragment containing a form for adding new source.

In 767e667046591851ee17f108a0fb6043579d916b, it has been changed to return JSON with just the `spouts`, which is almost identical to `GET /sources/spouts`.

Let’s switch to that and drop the old API endpoint.

Bumping API version to 8.0.0.