client: Use `currentTarget` instead of `target` for events
For bubbling events, the properties can have different values:
https://www.github.com/DefinitelyTyped/DefinitelyTyped/issues/11508#issuecomment-256045682

We typically intend to refer to the element the event handler is attached to. So `currentTarget` is the property we want:
https://developer.mozilla.org/en-US/docs/Web/API/Event/currentTarget#currenttarget_versus_target

The change is not necessary for `ChangeEvent` – we always attach such event handler to an input element, which cannot contain any children, so there is nothing that could resolve to a distinct `target`. But we will switch to `currentTarget` here as well for consistency:
https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/change_event

On the other hand, it is required for `SubmitEvent` DOM event. While, that event fires directly on `HTMLFormElement`, React maps it to `React.FormEvent` and does not specialize the `target` property for it.
https://developer.mozilla.org/en-US/docs/Web/API/SubmitEvent/SubmitEvent