Mark most classes final
Composition over inheritance as a best practice we should strive to follow:
https://en.wikipedia.org/wiki/Composition_over_inheritance

It should be fine to do this for controllers and daos since those are entirely internal. There is a slight chance that someone is extending one of the helpers in their spouts but I have not seen that on GitHub so I consider disallowing that acceptable.

We will still keep the following non-final:
- `daos\mysql\{Items,Sources,Statements,Tags}` because we were too lazy to follow the best practice there.
- `helpers\WebClient` because we mock it in tests.
- `spouts\spout` and its children because it would be too much of a BC break.

Let’s use search and replace and manually skip those that we do not want to finalize.