common: Print boot error on Configuration exception
Once we switch to Tracy, it would show a error 500 screen that would instruct user to set up a debug mode. But when the `Configuration` throws an exception, we have no way to determine the `Configuration->debug` property.

Let’s end early with a boot error instead of relying on Tracy catching the exception.

Nothing else in the file should throw either.